step3, InputView 클래스 파일 분리
