+++ src/main/kotlin/step3/racing/view/input/InputViewFactory.kt
@@ -1,0 +1,7 @@
+package step3.racing.view.input
+
+object InputViewFactory {
+    fun newInstance(): InputView {
+        return InputViewImpl()
+    }
+}
